+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# Version 26.9.11.2
+# Version 26.9.13.1
 
 Perla:
-- Overhauled Voice input UI
+- Added scoped commands section in Quick Actions
+- Added undo button for delete along with moving deleted files to system trash in the System Files view